+++ app/lib/screens/mess/factories/mess_item.dart
@@ -1,0 +1,9 @@
+class MessItem {
+  String name;
+
+  MessItem({required this.name});
+
+  factory MessItem.fromJson(Map<String, dynamic> json) {
+    return MessItem(name: json['name']);
+  }
+}
