+++ app/lib/screens/mess/mess.dart
@@ -1,13 +1,12 @@
 import 'dart:convert';
 
-import 'package:carousel_slider/carousel_slider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_flavor/flutter_flavor.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
-import 'package:ira/screens/mess/models.dart';
+import 'package:ira/screens/mess/components/mess_weekday.dart';
+import 'package:ira/screens/mess/factories/mess.dart';
 import 'package:ira/shared/app_scaffold.dart';
 import 'package:http/http.dart' as http;
-import '../../util/helpers.dart';
 
 class MessScreen extends StatefulWidget {
   const MessScreen({Key? key}) : super(key: key);
@@ -36,9 +35,7 @@ class _MessScreenState extends State<MessScreen> {
     Map<String, List<Mess>> mmp = {};
 
     if (response.statusCode == 200) {
-      // print('1');
       List decodedData = jsonDecode(response.body) as List;
-      // print(decodedData);
       mmp = {
         'data': decodedData.map<Mess>((json) => Mess.fromJson(json)).toList(),
       };
@@ -147,117 +144,5 @@ class _MessScreenState extends State<MessScreen> {
             ),
           );
         });
-  }
-}
-
-class MessWeekDay extends StatelessWidget {
-  final WeekDay weekDay;
-
-  const MessWeekDay({
-    Key? key,
-    required this.weekDay,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.all(20.0),
-      width: MediaQuery.of(context).size.width,
-      decoration: BoxDecoration(
-        color: Colors.white,
-        borderRadius: BorderRadius.circular(10.0),
-        boxShadow: const [
-          BoxShadow(
-              color: Color.fromARGB(255, 20, 12, 12),
-              blurRadius: 5.0,
-              offset: Offset(
-                0,
-                5,
-              ))
-        ],
-      ),
-      child: Column(children: [
-        Text(
-          weekDay.name,
-          style: const TextStyle(
-            fontSize: 25.0,
-            fontWeight: FontWeight.bold,
-          ),
-        ),
-        const SizedBox(
-          height: 10.0,
-        ),
-        SizedBox(
-          height: 450,
-          child: ListView.builder(
-            padding: EdgeInsets.zero,
-            scrollDirection: Axis.vertical,
-            shrinkWrap: true,
-            itemCount: weekDay.slots.length,
-            itemBuilder: (BuildContext context, int index) {
-              return MessSlot(slot: weekDay.slots[index]);
-            },
-          ),
-        ),
-      ]),
-    );
-  }
-}
-
-class MessSlot extends StatelessWidget {
-  final Slot slot;
-  const MessSlot({
-    Key? key,
-    required this.slot,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        Text(
-          slot.name,
-          style: const TextStyle(
-            fontSize: 16.0,
-          ),
-        ),
-        const SizedBox(
-          height: 5.0,
-        ),
-        Container(
-          padding: const EdgeInsets.all(10.0),
-          width: double.infinity,
-          decoration: BoxDecoration(
-            border: Border.all(color: Colors.black),
-            borderRadius: const BorderRadius.all(
-              Radius.circular(5.0),
-            ),
-          ),
-          child: Column(
-            // children: const [
-            //   Text('Veg Sevaiyaan'),
-            //   SizedBox(
-            //     height: 5.0,
-            //   ),
-            // ],
-            children: slot.items
-                .map(
-                  (e) => Column(
-                    children: [
-                      Text(e.name),
-                      const SizedBox(
-                        height: 5.0,
-                      )
-                    ],
-                  ),
-                )
-                .toList(),
-          ),
-        ),
-        const SizedBox(
-          height: 10.0,
-        ),
-      ],
-    );
   }
 }
