+++ app/pubspec.yaml
@@ -45,7 +45,6 @@ dependencies:
   localstorage: ^4.0.0+1
   qr_code_scanner: ^0.7.0
   flip_card: ^0.6.0
-  carousel_slider: ^4.1.1
 
 dev_dependencies:
   flutter_test:
